feat: controller 구현, refact : 4계층 아키텍처 리팩토링
