+++ product-service/src/main/java/com/bidulgi/productservice/domain/entity/constant/ProductStatus.java
@@ -1,0 +1,5 @@
+package com.bidulgi.productservice.domain.entity.constant;
+
+public enum ProductStatus {
+    PREPARE, ON_SALE, SOLD_OUT
+}
