--- product-service/src/main/java/com/bidulgi/productservice/entity/constant/ProductStatus.java
@@ -1,5 +1,0 @@
-package com.bidulgi.productservice.entity.constant;
-
-public enum ProductStatus {
-    PREPARE, ON_SALE, SOLD_OUT
-}
